fix: precision issue

+++ contracts/contracts/interfaces/IProviderRegistry.sol
@@ -34,7 +34,7 @@ interface IProviderRegistry {
     event PreconfManagerUpdated(address indexed newPreconfManager);
 
     /// @dev Event emitted when the fee percent is updated
-    event FeePercentUpdated(uint16 indexed newFeePercent);
+    event FeePercentUpdated(uint256 indexed newFeePercent);
 
     /// @dev Event emitted when there are insufficient funds to slash
     event InsufficientFundsToSlash(
